[fix] 목표 달성 판정을 기록 개수에서 ml 비례 크레딧으로 전환
887ml x 2잔 목표에서 200ml만 기록해도 50%(1잔)로 표시되던 문제 수정 - DayRecord.totalCount가
entries.size(기록 개수)만 세고 실제 ml은 무시하던 구조라, 소량을 여러 번 나눠 기록하면 목표를
손쉽게 달성할 수 있었다. cupSize 스냅샷을 추가해 totalCount를 ml 비례 크레딧으로 재계산하고,
AchievementChecker의 첫 잔 판정 등 연쇄 영향을 함께 정리했다.

+++ app/src/main/java/com/watering/app/core/model/DailyAchievement.kt
@@ -7,9 +7,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DailyAchievement(
     val dateKey: String,         // "yyyy-MM-dd" 형식
-    val totalCount: Int,
+    val totalCount: Double,      // DayRecord.totalCount(ml 비례 크레딧)를 그대로 복사
     val goal: Int
 ) {
-    val achievementRate: Double get() = if (goal == 0) 0.0 else totalCount.toDouble() / goal
+    val achievementRate: Double get() = if (goal == 0) 0.0 else totalCount / goal
     val isAchieved: Boolean get() = totalCount >= goal
 }
